+++ Extras/VHACD/src/vhacdMesh.cpp
@@ -47,7 +47,7 @@ double Mesh::ComputeVolume() const
 
     Vec3<double> ver0, ver1, ver2;
     double totalVolume = 0.0;
-    for (int t = 0; t < nT; t++) {
+    for (size_t t = 0; t < nT; t++) {
         const Vec3<int>& tri = GetTriangle(t);
         ver0 = GetPoint(tri[0]);
         ver1 = GetPoint(tri[1]);
@@ -115,7 +115,7 @@ bool Mesh::IsInside(const Vec3<double>& pt) const
     }
     Vec3<double> ver0, ver1, ver2;
     double volume;
-    for (int t = 0; t < nT; t++) {
+    for (size_t t = 0; t < nT; t++) {
         const Vec3<int>& tri = GetTriangle(t);
         ver0 = GetPoint(tri[0]);
         ver1 = GetPoint(tri[1]);
